% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_d.R
\name{cohens_d}
\alias{cohens_d}
\title{Calculate Cohen's d given some descriptive statistics}
\usage{
cohens_d(mean.one, mean.two, sd.one, sd.two, sd = c("same", "diff"))
}
\arguments{
\item{mean.one}{consumes the mean of first group}

\item{mean.two}{consumes the mean of second group}

\item{sd.one}{consumes the standard deviation of first group}

\item{sd.two}{consumes the standard deviation of second group. This can be left empty if the groups share the same standard deviation.}

\item{sd}{Type "same" if both groups have the same sd. Type "diff" if the groups do not have the same standard deviation. In this case, the pooled standard deviation is used.}
}
\value{
A string containing Cohen's d statistic that informs user of effect size
}
\description{
Calculate Cohen's d given some descriptive statistics
}
\examples{
cohens_d(0, 1.4, 0.6, sd = "same")
cohens_d(mean.one = 0, mean.two = 1, sd.one = 0.5, sd.two = 0.6, sd = "diff")
}
