% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_plot.R
\name{cohens_plot}
\alias{cohens_plot}
\title{Calculate and plots Cohen's d given given some descriptive statistic}
\usage{
cohens_plot(mean.one, mean.two, sd.one, sd.two, overlap = TRUE)
}
\arguments{
\item{mean.one}{consumes the mean of first group}

\item{mean.two}{consumes the mean of second group}

\item{sd.one}{consumes the standard deviation of first group}

\item{sd.two}{consumes the standard deviation of second group.}

\item{overlap}{default is set to true. Type false if you do not want an overlap in the plots.}
}
\value{
a ggplot showing the effect for the given mean and standard deviation
}
\description{
Calculate and plots Cohen's d given given some descriptive statistic
}
\examples{
cohens_plot(0, 1, 0.5, 0.8, overlap = FALSE)
cohens_plot(0, 1, 0.5, 0.8)
cohens_plot(0, 1, 0.5, 0.8, overlap = TRUE)
}
